AI算法/requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
memory-profiler==0.57.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
psutil==5.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
